refactor(checker): extract await-using context grammar into using_declaration_placement
core.rs was at the 2000-line arch-size cap, so the TS2853 gate could not land
there. Move the `await using` module/target/async-context family
(TS2852/TS2853/TS2854/TS1309) and its `enclosing_function_allows_await_using`
helper out of `check_variable_declaration_list_with_request` into
`using_declaration_placement.rs`, which already hosts the sibling TS1545-1548
placement grammar. Behavior is unchanged; core.rs drops to 1933 lines.
